Update node version 14.15.5 -> 15.14.0
This is the current v15 release and updating for v15 or above will allow
is to disable and possibly completely remove the
`NODEJS_CATCH_REJECTION` setting in emscripten.  This setting is a
workaround for the fact the older node version will return exit with 0
on unhandled promise rejection.

The downside to updating the version used by emsdk and therefore the
version against which we do all our emscripten testing is that we
potentially loose some coverage of compatability with older versions.
In practice I don't think we have seen a node compat issue in many
years.  The limits on the JS output produced by emscripten are all
related to targeting older browsers which tend to be a lot older than
any of the node versions we want to target.